backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
firebase-admin
openai
python-dotenv
pydantic
python-multipart